+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/CannonballBlock.java
@@ -1,0 +1,46 @@
+package net.mehvahdjukaar.supplementaries.common.block.blocks;
+
+import net.mehvahdjukaar.moonlight.api.block.WaterBlock;
+import net.mehvahdjukaar.supplementaries.common.block.ModBlockProperties;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import org.jetbrains.annotations.Nullable;
+
+public class CannonballBlock extends WaterBlock {
+
+    // jojo reference?
+    public static final IntegerProperty BALLS = ModBlockProperties.BALLS;
+
+    public CannonballBlock(BlockBehaviour.Properties properties) {
+        super(properties);
+        this.registerDefaultState((this.stateDefinition.any().setValue(BALLS, 1))
+                .setValue(WATERLOGGED, false));
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        super.createBlockStateDefinition(builder);
+        builder.add(BALLS);
+    }
+
+    @Override
+    public boolean canBeReplaced(BlockState state, BlockPlaceContext useContext) {
+        return !useContext.isSecondaryUseActive() && useContext.getItemInHand().is(this.asItem()) && state.getValue(BALLS) < 4 ||
+                super.canBeReplaced(state, useContext);
+    }
+
+    @Nullable
+    @Override
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        BlockState blockState = context.getLevel().getBlockState(context.getClickedPos());
+        if (blockState.is(this)) {
+            return blockState.setValue(BALLS, Math.min(4, blockState.getValue(BALLS) + 1));
+        } else {
+            return super.getStateForPlacement(context);
+        }
+    }
+}