+++ fabric/src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/fabric/QuiverArrowSelectGuiImpl.java
@@ -1,16 +1,11 @@
 package net.mehvahdjukaar.supplementaries.client.renderers.fabric;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.PoseStack;
-import me.shedaniel.clothconfig2.ClothConfigDemo;
-import net.mehvahdjukaar.moonlight.api.platform.configs.fabric.FabricConfigSpec;
 import net.mehvahdjukaar.supplementaries.client.QuiverArrowSelectGui;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
-import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.item.ItemStack;
@@ -39,7 +34,7 @@ public class QuiverArrowSelectGuiImpl extends QuiverArrowSelectGui {
 
         RenderSystem.enableBlend();
         RenderSystem.defaultBlendFunc();
-        graphics.fill( nx - 2, ny - 2, nx + fontWidth + 2, ny + 9 + 2, this.minecraft.options.getBackgroundColor(0));
+        graphics.fill(nx - 2, ny - 2, nx + fontWidth + 2, ny + 9 + 2, this.minecraft.options.getBackgroundColor(0));
         Font font = this.getFont();
         nx = (screenWidth - font.width(highlightTip)) / 2;
         graphics.drawString(font, highlightTip, nx, ny, 0xFFFFFF + (l << 24));
@@ -48,10 +43,11 @@ public class QuiverArrowSelectGuiImpl extends QuiverArrowSelectGui {
 
 
     public void render(GuiGraphics graphics, float partialTicks) {
-        if(isActive()) {
+        if (isActive()) {
+            var w = this.minecraft.getWindow();
             renderQuiverContent(graphics, partialTicks,
-                    Minecraft.getInstance().getWindow().getGuiScaledWidth(),
-                    Minecraft.getInstance().getWindow().getGuiScaledHeight());
+                    w.getGuiScaledWidth(),
+                    w.getGuiScaledHeight());
         }
     }